requirements.txt
# ⚠️ IMPORTANT: Python 3.13 is NOT compatible with TensorFlow
# You MUST use Python 3.11 or earlier to run this project
#
# Install Python 3.11:
#   macOS: brew install python@3.11
#   Or download from: https://www.python.org/downloads/release/python-31110/
#
# Create virtual environment with Python 3.11:
#   python3.11 -m venv venv
#   source venv/bin/activate
#
# Then install these packages:

tensorflow>=2.15.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
